+++ src/com/nixinova/mineo/world/blocks/Block.java
@@ -1,15 +1,12 @@
 package com.nixinova.mineo.world.blocks;
 
-import com.nixinova.mineo.ui.graphics.Render;
-import com.nixinova.mineo.ui.graphics.Texture;
-
 public class Block {
 	public static final Block AIR = new Block(null);
 	public static final Block BEDROCK = new Block("blocks/bedrock");
 	public static final Block STONE = new Block("blocks/stone");
 	public static final Block DIRT = new Block("blocks/dirt");
-	public static final Block GRASS = new Block("blocks/grass");
-	public static final Block LOG = new Block("blocks/log");
+	public static final Block GRASS = new Block("blocks/grass_top", "blocks/grass_top", "blocks/dirt");
+	public static final Block LOG = new Block("blocks/log_top", "blocks/log_side", "blocks/log_top");
 	public static final Block LEAF = new Block("blocks/leaf");
 
 	public static final Block DEBUG = new Block("debug");
@@ -23,20 +20,51 @@ public class Block {
 		GRASS,
 	};
 
-	private Render texture;
-	private String name;
-
-	public Block(String texture) {
-		this.texture = texture == null ? null : Texture.loadTexture(texture);
-		this.name = texture;
+	public enum Side {
+		TOP,
+		SIDE,
+		BOTTOM,
 	}
 
-	public Render getTexture() {
-		return this.texture;
+	private BlockTexture topTexture, sideTexture, bottomTexture;
+
+	public Block(String textureName) {
+		var texture = textureName == null ? null : new BlockTexture(textureName);
+		this.topTexture = this.sideTexture = this.bottomTexture = texture;
+	}
+	
+	public Block(String topTexture, String sideTexture, String bottomTexture) {
+		this.topTexture = new BlockTexture(topTexture);
+		this.sideTexture = new BlockTexture(sideTexture);
+		this.bottomTexture = new BlockTexture(bottomTexture);
+	}
+	
+	public BlockTexture getTexture() {
+		// default to the side texture
+		return this.sideTexture;
 	}
 
-	public String getTextureName() {
-		return this.name;
+	public BlockTexture getTexture(Side side) {
+		switch (side) {
+			case TOP:
+				return this.topTexture;
+			case SIDE:
+				return this.sideTexture;
+			case BOTTOM:
+				return this.bottomTexture;
+			default:
+				return null;
+		}
 	}
-
+	
+	public BlockTexture getTexture(BlockFace face) {
+		switch (face) {
+			case YMAX:
+				return this.topTexture;
+			case YMIN:
+				return this.bottomTexture;
+			default:
+				return this.sideTexture;
+		}
+	}
 }